Add files via upload
Your project is to read in a series of fractions from a text file, which will have each line formatted as follows: “A/B”.  A sample text file is listed below, and the purpose of your program is to read in each fraction and count the number of occurrences for the current fraction.  When all the input is consumed (or as the input is consumed), your program will print out its list of unique fraction and their corresponding count — see the output below (and you may assume no blank lines or misleading characters; see the text file link on the website for one of the actual inputs I’ll use when testing your submission).

The fractions should always be represented as two integers - DO NOT USE DECIMALS.

You do not need to reduce fractions for this assignment. We will discuss in class how to compare values.